Implement combined relative evaluation heuristic

+++ src/logic/ai/evaluation/Evaluator.kt
@@ -7,5 +7,5 @@ import logic.board.GameBoard
  * Created by r.makowiecki on 16/05/2017.
  */
 interface Evaluator {
-    fun evaluate(board: GameBoard, ownedFieldsType: FieldState): Int
+    fun evaluate(board: GameBoard, ownedFieldsType: FieldState): Float
 }
