+++ src/logic/ai/evaluation/field_weights/FieldWeightProvider.kt
@@ -4,5 +4,5 @@ package logic.ai.evaluation.field_weights
  * Created by r.makowiecki on 17/05/2017.
  */
 interface FieldWeightProvider {
-    fun getFieldWeights() : Array<Int>
+    fun getFieldWeights() : Array<Float>
 }
